+++ src/router.js
@@ -97,15 +97,15 @@ let router = new Router({
 
 // Used to check if a route is projected, if it is then you are required to
 // be authenticated before you can access it
-router.beforeEach((to, from, next) => {
-	if (to.matched.some(record => record.meta.requiresAuth)) {
-		if (store.getters.isLoggedIn) {
-			next();
-			return;
-		}
-		next("/login");
-	} else {
-		next();
-	}
-});
+// router.beforeEach((to, from, next) => {
+// 	if (to.matched.some(record => record.meta.requiresAuth)) {
+// 		if (store.getters.isLoggedIn) {
+// 			next();
+// 			return;
+// 		}
+// 		next("/login");
+// 	} else {
+// 		next();
+// 	}
+// });
 export default router;